Bug fix
Had an issue where the audio file for 'Lands of Always Winter' wasn't playing on the Pages link but worked locally. Found out that this was due Local file systems (like macOS) are often case-insensitive, while GitHub Pages (which is hosted on a Linux server) is case-sensitive. For example, Music/Winter.mp3 and music/winter.mp3 are treated as different paths on GitHub Pages. I didn't take this into consideration and will now commit to see if this change fixes the issue.